requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-flex-fields==1.0.2
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
